extract "hard logout" clearing data to the new use case
